fix isLocal again! withdraw Grib2Tables.getParameter(disc, cat, param) from public API, in preparation of more complex matching. (part 1).
